+++ led_matrix_battery/led_matrix/display/patterns.py
@@ -8,8 +8,8 @@ It includes functions for displaying checkerboards, gradients, and other visual 
 import serial
 import time
 
-from .constants import WIDTH, HEIGHT
-from .hardware import (
+from ..constants import WIDTH, HEIGHT
+from ..hardware import (
     CommandVals, PatternVals, send_command, send_serial,
     brightness, get_status, set_status
 )
@@ -65,7 +65,7 @@ def pattern(dev, p):
 
 def send_col(dev, s, x, vals):
     """Stage greyscale values for a single column. Must be committed with commit_cols()"""
-    from .hardware import FWK_MAGIC
+    from ..hardware import FWK_MAGIC
     command = FWK_MAGIC + [CommandVals.StageGreyCol, x] + vals
     send_serial(dev, s, command)
 
@@ -73,7 +73,7 @@ def send_col(dev, s, x, vals):
 def commit_cols(dev, s):
     """Commit the changes from sending individual cols with send_col(), displaying the matrix.
     This makes sure that the matrix isn't partially updated."""
-    from .hardware import FWK_MAGIC
+    from ..hardware import FWK_MAGIC
     command = FWK_MAGIC + [CommandVals.DrawGreyColBuffer, 0x00]
     send_serial(dev, s, command)